requirements.txt
# HTTP客户端
requests>=2.31.0
aiohttp>=3.9.1

# API客户端
openai>=1.0.0
google-api-python-client>=2.108.0
youtube_transcript_api>=0.6.1

# 环境变量
python-dotenv>=1.0.0

# 类型提示
typing>=3.7.4.3

# 内容解析
beautifulsoup4>=4.12.3
lxml>=5.1.0

# 日志和异步
logging>=0.5.1.2
asyncio>=3.4.3

# 工具库
uuid>=1.30
json>=2.0.9